+++ .eslintrc.js
@@ -4,7 +4,12 @@ module.exports = {
   env: {
     node: true,
   },
-  extends: ['eslint:recommended', 'plugin:jest-formatting/recommended'],
+  extends: [
+    'eslint:recommended',
+    'plugin:jest-formatting/recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+  ],
   parserOptions: {
     ecmaVersion: 2020,
   },
